+++ app/page.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import { DevToolsWorkbench } from "./dev-tools-workbench";
+import { PwaClient } from "./pwa-client";
 
 export default function Home() {
-  return <DevToolsWorkbench />;
+  return <><DevToolsWorkbench /><PwaClient /></>;
 }